+++ src/plone/app/imagecropping/browser/scalingdx.py
@@ -25,5 +25,9 @@ class ImageScalingDX(ScalingOverrides, NFImageScaling):
     def scale(self, fieldname=None, scale=None, height=None, width=None,
               direction='thumbnail', **parameters):
         self._need_rescale(fieldname, scale)
+        # if direction is 'down' and we have a cropped scale
+        # deliver it instead of standard 'down' scale
+        if direction == 'down' and not self._allow_rescale:
+            direction = 'thumbnail'
         return super(ImageScalingDX, self).scale(
             fieldname, scale, height, width, direction, **parameters)